requirements.txt
pyserial>=3.5
websockets>=12.0
pyserial>=3.5
